Created fuction to search for specific objects: Products and Logins
