+++ app/controllers/Application.scala
@@ -1,16 +1,16 @@
 package controllers
 
-import models.{LoginList, ProductList}
+import models.{Login, Product}
 import play.api.mvc._
 
 class Application extends Controller {
 
   def diddly = Action {
-    var products = new ProductList
-    products.generate()
-    var logins = new LoginList
-    logins.generate()
-    Ok(views.html.test(products, logins))
+
+    Product.generate()
+    Login.generate()
+
+    Ok(views.html.test())
   }
 
 }
